requirements.txt
numpy
pandas
h5py
astropy
matplotlib
seaborn
tqdm
Pillow
torch
torchvision
timm